move entity_localtor to its own file
