+++ field_friend/automations/__init__.py
@@ -1,13 +1,14 @@
 
 from .automation_watcher import AutomationWatcher
 from .battery_watcher import BatteryWatcher
+from .entity_locator import EntityLocator
 from .field import Field, Row, RowSupportPoint
 from .field_provider import FieldProvider
 from .implements.implement import Implement
 from .kpi_provider import KpiProvider
 from .path_provider import Path, PathProvider
 from .plant import Plant
-from .plant_locator import EntityLocator, PlantLocator
+from .plant_locator import PlantLocator
 from .plant_provider import PlantProvider
 from .puncher import Puncher
