+++ field_friend/automations/plant_locator.py
@@ -9,6 +9,7 @@ from nicegui import ui
 from rosys.vision import Autoupload
 
 from ..vision.zedxmini_camera import StereoCamera
+from .entity_locator import EntityLocator
 from .plant import Plant
 
 WEED_CATEGORY_NAME = ['coin', 'weed', 'weedy_area', ]
@@ -25,38 +26,7 @@ class DetectorError(Exception):
     pass
 
 
-class EntityLocator(rosys.persistence.PersistentModule):
-    def __init__(self, system: System) -> None:
-        super().__init__()
-        self.log = logging.getLogger('field_friend.entity_locator')
-        self.system = system
-
-        self.is_paused = True
-
-    def pause(self) -> None:
-        if self.is_paused:
-            return
-        self.log.info('pausing locator')
-        self.is_paused = True
-
-    def resume(self) -> None:
-        if not self.is_paused:
-            return
-        self.log.info('resuming locator')
-        self.is_paused = False
-
-    def developer_ui(self) -> None:
-        pass
-
-    def backup(self) -> dict:
-        return {}
-
-    def restore(self, data: dict[str, Any]) -> None:
-        pass
-
-
 class PlantLocator(EntityLocator):
-
     def __init__(self, system: System) -> None:
         super().__init__(system)
         self.log = logging.getLogger('field_friend.plant_locator')
